backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.1
Deprecated==1.2.13
Django==4.0.5
djangorestframework==3.13.1
Faker==13.12.1
flake8==4.0.1
iniconfig==1.1.1
kombu==5.2.4
mccabe==0.6.1
packaging==21.3
pika==1.2.1
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
